Move mockData to a higher level so it removes duplicated code
